Split reference core and SQL patterns
